Working on the realtime-functionalities.
I tried starting to include the backend-model classes, but it has so far been unsuccessful.
(i) You can't yet create room-Instances, as that requires the GameObjectService-class, which requires the ObjectService class, which does not yet exist
(ii) Creating a Participant-Instance in the ParticipantController-Constructor does not work, as apparently, 'new Participant' is not a constructor at that time (???)

Besides that, I made the TypeChecker-class an exportable module and also added some small changes here and there to make some classes work along my implementation of the realtime-Communication.

Still To-Do:
(i) Movement
(ii) Figure out if no new avatar is being drawn because of the view-class of because i fucked up
(iii) Alter backend-model to work along rt-func