Added client-side handling of allchat. allchat is now properly displayed and loaded on room-entry. On switching rooms, allchat is emptied.
To-Do
Allchat in Foodcourt does not work - might be because server gets the messages from the room it assumes the participant to be in?
HTML may still be a bit wonky - each allchat-message is a div inside the large allchat-div. Also using jQuery to add messages to display. This can maybe be done more smoothly?